pyproject.toml
[project]
name = "chain-of-extraction"
version = "0.1.0"
description = "Automated pipeline for OCR and LLM-based extraction from cosmetic regulation PDFs"
authors = [
    {name = "Chain of Extraction Team"}
]
dependencies = [
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
]
requires-python = ">=3.8"

[project.scripts]
chain-extract = "chain_of_extraction.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"